+++ crates/ruma-events/src/secret_storage/default_key.rs
@@ -12,5 +12,16 @@ use crate::macros::EventContent;
 #[ruma_event(type = "m.secret_storage.default_key", kind = GlobalAccountData)]
 pub struct SecretStorageDefaultKeyEventContent {
     /// The ID of the default key.
-    pub key: String,
+    #[serde(rename = "key")]
+    pub key_id: String,
+}
+
+impl SecretStorageDefaultKeyEventContent {
+    /// Create a new [`SecretStorageDefaultKeyEventContent`] with the given key ID.
+    ///
+    /// Uploading this to the account data will mark the secret storage key with the given key ID as
+    /// the default key.
+    pub fn new(key_id: String) -> Self {
+        Self { key_id }
+    }
 }