+++ crates/ruma-events/CHANGELOG.md
@@ -9,6 +9,8 @@ Bug fixes:
 
 Breaking changes:
 
+- Rename the `key` field in the `SecretStorageDefaultKeyEventContent` to
+  `key_id`.
 - Make `in_reply_to` field of `Thread` optional
   - It was wrong to be mandatory, spec was unclear (clarified [here](https://github.com/matrix-org/matrix-spec/pull/1439))
 - Remove `AnswerSessionDescription` and `OfferSessionDescription` types, use `SessionDescription`